refactor: 토론게시판에서만 투표가 있으면 삭제하도록 수정

+++ backend/src/main/java/com/woowacourse/gongseek/article/application/ArticleService.java
@@ -176,6 +176,9 @@ public class ArticleService {
     }
 
     private void deleteVoteHistory(Article article) {
+        if (article.getCategory().equals(Category.QUESTION)) {
+            return;
+        }
         Vote vote = voteRepository.findByArticleId(article.getId())
                 .orElseThrow(() -> new VoteNotFoundException(article.getId()));
         voteHistoryRepository.deleteAllByVoteItemIn(voteItemRepository.findAllByVoteId(vote.getId()));
